server/requirements.txt
asgiref==3.7.2
channels==4.0.0
dj_rest_auth==5.0.1
Django==4.1.7
django_allauth==0.54.0
djangorestframework==3.14.0
docx==0.2.4
Faker==20.1.0
langchain==0.0.348
openai==1.3.8
pinecone_client==2.2.4
protobuf==4.25.1
python-dotenv==1.0.0
Requests==2.31.0
torch==2.1.1
sentence-transformers
text_generation
python-docx
daphne
google-generativeai
drf_yasg